requirements.txt
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cryptography==3.0
cycler==0.10.0
decorator==4.4.2
future==0.18.2
idna==2.10
kaleido==0.0.3
kiwisolver==1.2.0
matplotlib==3.3.0
numpy==1.19.1
Pillow==7.2.0
plotly==4.9.0
pycparser==2.20
pydantic==1.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
python-telegram-bot==12.8
PyYAML==5.3.1
requests==2.24.0
retrying==1.3.3
six==1.15.0
tabulate==0.8.7
tornado==6.0.4
urllib3==1.25.10
